Reorganize namespace and file location
