+++ Medidata.Rave.Tsdv.Loader.Sample/Program.cs
@@ -3,6 +3,7 @@ using System.IO;
 using System.Linq;
 using Medidata.Cloud.ExcelLoader.Helpers;
 using Medidata.Cloud.ExcelLoader.SheetDefinitions;
+using Medidata.Rave.Tsdv.Loader.SheetDefinitions;
 using Medidata.Rave.Tsdv.Loader.SheetDefinitions.v1;
 using Microsoft.Practices.Unity;
 using Microsoft.Practices.Unity.Configuration;
