+++ Medidata.Rave.Tsdv.Loader.Sample/unity.config
@@ -5,7 +5,7 @@
     <register type="Medidata.Interfaces.Localization.ILocalization, Medidata.Interfaces"
               mapTo="Medidata.Rave.Tsdv.Loader.Sample.Localization, Medidata.Rave.Tsdv.Loader.Sample" />
     <register type="Medidata.Rave.Tsdv.Loader.ITsdvExcelLoaderFactory, Medidata.Rave.Tsdv.Loader"
-              mapTo="Medidata.Rave.Tsdv.Loader.TsdvLoaderV1Factory, Medidata.Rave.Tsdv.Loader" >
+              mapTo="Medidata.Rave.Tsdv.Loader.SheetDefinitions.v1.TsdvLoaderFactory, Medidata.Rave.Tsdv.Loader" >
       <lifetime type="singleton" />
     </register>
   </container>
