+++ Medidata.Rave.Tsdv.Loader.Tests/Medidata.Rave.Tsdv.Loader.Tests.csproj
@@ -75,8 +75,8 @@
   </Choose>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TsdvLoaderV1FactoryTests.cs" />
-    <Compile Include="TsdvPresentationLoaderFactoryTests.cs" />
+    <Compile Include="SheetDefinitions\v1\TsdvLoaderFactoryTests.cs" />
+    <Compile Include="SheetDefinitions\Presentation\TsdvPresentationLoaderFactoryTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
